+++ crates/rerun/src/lib.rs
@@ -5,7 +5,7 @@
 //! There is also a `rerun` binary.
 //! The binary is required in order to stream log data
 //! over the networks, and to open our `.rrd` data files.
-//! If you need it, install the `rerun` binary with `cargo install rerun`.
+//! If you need it, install the `rerun` binary with `cargo install rerun-cli`.
 //!
 //! ## Feature flags
 #![doc = document_features::document_features!()]
