+++ examples/rust/minimal_options/Cargo.toml
@@ -7,7 +7,10 @@ license = "MIT OR Apache-2.0"
 publish = false
 
 [dependencies]
-rerun = { path = "../../../crates/rerun", features = ["web_viewer"] }
+rerun = { path = "../../../crates/rerun", features = [
+  "native_viewer",
+  "web_viewer",
+] }
 
 anyhow = "1.0"
 clap = { version = "4.0", features = ["derive"] }
