+++ examples/rust/objectron/Cargo.toml
@@ -8,7 +8,10 @@ publish = false
 
 
 [dependencies]
-rerun = { path = "../../../crates/rerun", features = ["web_viewer"] }
+rerun = { path = "../../../crates/rerun", features = [
+  "native_viewer",
+  "web_viewer",
+] }
 
 anyhow = "1.0"
 clap = { version = "4.0", features = ["derive"] }
